Fix some bugs and add `TestLogicGate`

+++ StoneRed.LogicSimulator/WorldSaveSystem/WorldData.cs
@@ -8,11 +8,11 @@ internal class WorldData
 {
     public IEnumerable<LogicGate> LogicGates { get; set; }
 
-    public int SaveVersion { get; set; }
+    public ushort SaveVersion { get; set; }
 
     public string SaveName { get; set; }
 
-    public WorldData(IEnumerable<LogicGate> logicGates, int saveVersion, string saveName)
+    public WorldData(IEnumerable<LogicGate> logicGates, ushort saveVersion, string saveName)
     {
         LogicGates = logicGates;
         SaveVersion = saveVersion;
